+++ warp/render/render_usd.py
@@ -13,9 +13,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from __future__ import annotations
+
 import numpy as np
 
 import warp as wp
+
+UP_AXIS_TOKEN = ("X", "Y", "Z")
 
 
 def _usd_add_xform(prim):
@@ -158,12 +162,13 @@ class UsdRenderer:
     def add_shape_instance(
         self,
         name: str,
-        shape,
+        shape: int,
         body,
         pos: tuple,
         rot: tuple,
         scale: tuple = (1.0, 1.0, 1.0),
-        color: tuple = (1.0, 1.0, 1.0),
+        color1=None,
+        color2=None,
         custom_index: int = -1,
         visible: bool = True,
     ):
@@ -186,9 +191,13 @@ class UsdRenderer:
         rot: tuple,
         width: float,
         length: float,
-        color: tuple = None,
+        color: tuple = (1.0, 1.0, 1.0),
+        color2=None,
         parent_body: str = None,
         is_template: bool = False,
+        u_scaling=1.0,
+        v_scaling=1.0,
+        visible: bool = True,
     ):
         """
         Render a plane with the given dimensions.
@@ -239,6 +248,7 @@ class UsdRenderer:
         if not is_template:
             _usd_set_xform(plane, pos, rot, (1.0, 1.0, 1.0), 0.0)
 
+        plane.GetVisibilityAttr().Set("inherited" if visible else "invisible", self.time)
         return prim_path
 
     def render_ground(self, size: float = 100.0, plane=None):
@@ -290,6 +300,7 @@ class UsdRenderer:
         parent_body: str = None,
         is_template: bool = False,
         color: tuple = None,
+        visible: bool = True,
     ):
         """Debug helper to add a sphere for visualization
 
@@ -328,6 +339,7 @@ class UsdRenderer:
         if not is_template:
             _usd_set_xform(sphere, pos, rot, (1.0, 1.0, 1.0), self.time)
 
+        sphere.GetVisibilityAttr().Set("inherited" if visible else "invisible", self.time)
         return prim_path
 
     def render_capsule(
@@ -339,7 +351,9 @@ class UsdRenderer:
         half_height: float,
         parent_body: str = None,
         is_template: bool = False,
+        up_axis: int = 1,
         color: tuple = None,
+        visible: bool = True,
     ):
         """
         Debug helper to add a capsule for visualization
@@ -372,7 +386,7 @@ class UsdRenderer:
 
         capsule.GetRadiusAttr().Set(float(radius))
         capsule.GetHeightAttr().Set(float(half_height * 2.0))
-        capsule.GetAxisAttr().Set("Y")
+        capsule.GetAxisAttr().Set(UP_AXIS_TOKEN[up_axis])
 
         if color is not None:
             capsule.GetDisplayColorAttr().Set([Gf.Vec3f(color)], self.time)
@@ -382,6 +396,7 @@ class UsdRenderer:
         if not is_template:
             _usd_set_xform(capsule, pos, rot, (1.0, 1.0, 1.0), self.time)
 
+        capsule.GetVisibilityAttr().Set("inherited" if visible else "invisible", self.time)
         return prim_path
 
     def render_cylinder(
@@ -393,7 +408,9 @@ class UsdRenderer:
         half_height: float,
         parent_body: str = None,
         is_template: bool = False,
+        up_axis: int = 1,
         color: tuple = None,
+        visible: bool = True,
     ):
         """
         Debug helper to add a cylinder for visualization
@@ -426,7 +443,7 @@ class UsdRenderer:
 
         cylinder.GetRadiusAttr().Set(float(radius))
         cylinder.GetHeightAttr().Set(float(half_height * 2.0))
-        cylinder.GetAxisAttr().Set("Y")
+        cylinder.GetAxisAttr().Set(UP_AXIS_TOKEN[up_axis])
 
         if color is not None:
             cylinder.GetDisplayColorAttr().Set([Gf.Vec3f(color)], self.time)
@@ -436,6 +453,7 @@ class UsdRenderer:
         if not is_template:
             _usd_set_xform(cylinder, pos, rot, (1.0, 1.0, 1.0), self.time)
 
+        cylinder.GetVisibilityAttr().Set("inherited" if visible else "invisible", self.time)
         return prim_path
 
     def render_cone(
@@ -447,7 +465,9 @@ class UsdRenderer:
         half_height: float,
         parent_body: str = None,
         is_template: bool = False,
+        up_axis: int = 1,
         color: tuple = None,
+        visible: bool = True,
     ):
         """
         Debug helper to add a cone for visualization
@@ -480,7 +500,7 @@ class UsdRenderer:
 
         cone.GetRadiusAttr().Set(float(radius))
         cone.GetHeightAttr().Set(float(half_height * 2.0))
-        cone.GetAxisAttr().Set("Y")
+        cone.GetAxisAttr().Set(UP_AXIS_TOKEN[up_axis])
 
         if color is not None:
             cone.GetDisplayColorAttr().Set([Gf.Vec3f(color)], self.time)
@@ -490,6 +510,7 @@ class UsdRenderer:
         if not is_template:
             _usd_set_xform(cone, pos, rot, (1.0, 1.0, 1.0), self.time)
 
+        cone.GetVisibilityAttr().Set("inherited" if visible else "invisible", self.time)
         return prim_path
 
     def render_box(
@@ -501,6 +522,7 @@ class UsdRenderer:
         parent_body: str = None,
         is_template: bool = False,
         color: tuple = None,
+        visible: bool = True,
     ):
         """Debug helper to add a box for visualization
 
@@ -538,6 +560,7 @@ class UsdRenderer:
         if not is_template:
             _usd_set_xform(cube, pos, rot, extents, self.time)
 
+        cube.GetVisibilityAttr().Set("inherited" if visible else "invisible", self.time)
         return prim_path
 
     def render_ref(self, name: str, path: str, pos: tuple, rot: tuple, scale: tuple, color: tuple = None):
@@ -573,6 +596,8 @@ class UsdRenderer:
         update_topology=False,
         parent_body: str = None,
         is_template: bool = False,
+        smooth_shading: bool = True,
+        visible: bool = True,
     ):
         from pxr import Sdf, UsdGeom
 
@@ -620,9 +645,18 @@ class UsdRenderer:
         if not is_template:
             _usd_set_xform(mesh, pos, rot, scale, self.time)
 
+        mesh.GetVisibilityAttr().Set("inherited" if visible else "invisible", self.time)
         return prim_path
 
-    def render_line_list(self, name, vertices, indices, color, radius):
+    def render_line_list(
+        self,
+        name: str,
+        vertices,
+        indices,
+        color: tuple = None,
+        radius: float = 0.01,
+        visible: bool = True,
+    ):
         """Debug helper to add a line list as a set of capsules
 
         Args:
@@ -672,9 +706,11 @@ class UsdRenderer:
         instancer.GetScalesAttr().Set(line_scales, self.time)
         instancer.GetProtoIndicesAttr().Set([0] * num_lines, self.time)
 
-    #      instancer.GetPrimvar("displayColor").Set(line_colors, time)
+        # instancer.GetPrimvar("displayColor").Set(line_colors, time)
 
-    def render_line_strip(self, name: str, vertices, color: tuple, radius: float = 0.01):
+        instancer.GetVisibilityAttr().Set("inherited" if visible else "invisible", self.time)
+
+    def render_line_strip(self, name: str, vertices, color: tuple = None, radius: float = 0.01, visible: bool = True):
         from pxr import Gf, UsdGeom
 
         num_lines = int(len(vertices) - 1)
@@ -717,7 +753,9 @@ class UsdRenderer:
         instancer_capsule = UsdGeom.Capsule.Get(self.stage, instancer.GetPath().AppendChild("capsule"))
         instancer_capsule.GetDisplayColorAttr().Set([Gf.Vec3f(color)], self.time)
 
-    def render_points(self, name: str, points, radius, colors=None):
+        instancer.GetVisibilityAttr().Set("inherited" if visible else "invisible", self.time)
+
+    def render_points(self, name: str, points, radius, colors=None, visible: bool = True):
         from pxr import Gf, UsdGeom
 
         instancer_path = self.root.GetPath().AppendChild(name)
@@ -755,6 +793,8 @@ class UsdRenderer:
         else:
             instancer.GetPointsAttr().Set(points, self.time)
             instancer.GetDisplayColorAttr().Set(colors, self.time)
+
+        instancer.GetVisibilityAttr().Set("inherited" if visible else "invisible", self.time)
 
     def update_body_transforms(self, body_q):
         from pxr import Sdf, UsdGeom