+++ CHANGELOG.md
@@ -10,6 +10,8 @@
   setting or the module-level `"compile_time_trace"` option. When used, JSON files in the Trace Event
   format will be written in the kernel cache, which can be opened in a viewer like `chrome://tracing/`
   ([GH-609](https://github.com/NVIDIA/warp/issues/609)).
+- Add support for animating visibility of objects in the USD renderer
+  ([GH-598](https://github.com/NVIDIA/warp/issues/598)).
 
 ### Changed